requirement.txt
faiss-cpu
numpy
scikit-learn
umap-learn
annoy
scann
matplotlib
scipy
pandas
